Split ph-events plugin into component files, improve style, fix filtering issue when deselecting filters, fix date math / timezone issue on date range filters
