+++ wordpress/protohaven-events/src/view.js
@@ -1,10 +1,9 @@
-import { App } from './lib';
+import { App } from './app';
 import { createRoot } from 'react-dom/client';
 
 window.addEventListener("load", (event) => {
 	const elem = document.getElementById("protohaven-events");
 	const data = JSON.parse(elem.children[0].innerHTML);
-	console.log("initial data", data);
 	const root = createRoot(elem);
 	root.render(<App initialData={data} {...elem.dataset} />);
 });
